Sign app packages
It turns out this is totally unnecessary for the purposes of running and
debugging, which is all I will care about for the foreseeable future. So
I'm going to take a wrecking ball to most of this stuff in the next
commit. Kept for posterity, and in case I eventually want to support
msix distribution.