Copy pdb to package directory
This will cause the VS to pick up the copied pdb, enabling us to link a
new exe while the old one momentarily continues to run in the debugger.

In theory, this fixes the last of the bugs where a project fails to
build if a previous build is running.